+++ packages/swapkit/types/src/network.ts
@@ -209,7 +209,7 @@ export const EVMChainList: EVMChain[] = [
   Chain.Avalanche,
   Chain.BinanceSmartChain,
   Chain.Arbitrum,
-  // Chain.Base,
+  Chain.Base,
   Chain.Optimism,
   Chain.Polygon,
 ];