+++ packages/wallets/wc/src/constants.ts
@@ -87,15 +87,16 @@ export enum DEFAULT_NEAR_METHODS {
 export enum DEFAULT_NEAR_EVENTS {}
 
 export const WC_SUPPORTED_CHAINS = [
-  Chain.Binance, // Not supported by Wallet Connect
-  Chain.BinanceSmartChain,
+  // Chain.Binance, // Not supported by Wallet Connect
+  // Chain.BinanceSmartChain,
   Chain.Ethereum,
-  Chain.THORChain,
+  // Chain.THORChain,
   Chain.Avalanche,
-  Chain.Arbitrum,
+  // Chain.Arbitrum,
   Chain.Optimism,
   Chain.Polygon,
-  Chain.Cosmos,
-  Chain.Maya,
-  Chain.Kujira,
+  Chain.Base,
+  // Chain.Cosmos,
+  // Chain.Maya,
+  // Chain.Kujira,
 ] as const;