+++ e2e/wallets/intergration-db/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @coinmasters/integration-coins
 
+## 1.7.4
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-db@4.7.4
+  - @coinmasters/types@4.7.4
+  - @coinmasters/core@11.0.4
+
 ## 1.7.3
 
 ### Patch Changes