+++ packages/swapkit/core/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @coinmasters/core
 
+## 11.0.4
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/toolbox-cosmos@11.0.4
+  - @coinmasters/helpers@11.0.4
+  - @coinmasters/tokens@3.7.4
+  - @coinmasters/toolbox-utxo@11.0.4
+  - @coinmasters/types@4.7.4
+  - @coinmasters/toolbox-evm@11.0.4
+  - @coinmasters/api@3.7.4
+
 ## 11.0.3
 
 ### Patch Changes