+++ packages/swapkit/entities/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @thorswap-lib/swapkit-entities
 
+## 3.7.4
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.7.4
+
 ## 3.7.3
 
 ### Patch Changes