+++ packages/swapkit/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @swapkit/tokens
 
+## 3.7.4
+
+### Patch Changes
+
+- bump
+
 ## 3.7.3
 
 ### Patch Changes