+++ packages/toolboxes/cosmos/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @coinmasters/toolbox-cosmos
 
+## 11.0.4
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/helpers@11.0.4
+  - @coinmasters/types@4.7.4
+  - @coinmasters/api@3.7.4
+
 ## 11.0.3
 
 ### Patch Changes