+++ packages/wallets/evm-extensions/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @coinmasters/wallet-evm-extensions
 
+## 11.0.4
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.7.4
+  - @coinmasters/toolbox-evm@11.0.4
+
 ## 11.0.3
 
 ### Patch Changes