+++ packages/wallets/keplr/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @coinmasters/wallet-keplr
 
+## 11.0.4
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/toolbox-cosmos@11.0.4
+  - @coinmasters/helpers@11.0.4
+  - @coinmasters/types@4.7.4
+
 ## 11.0.3
 
 ### Patch Changes